backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.1.1
pydantic-settings>=2.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
requests>=2.31.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.19
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
google-auth>=2.0.0
email-validator>=2.0.0
httpx>=0.24.0
slowapi>=0.1.9
python-multipart>=0.0.6
psycopg2-binary>=2.9.0
